generateExpressionArithmetic gets datatypes from the  stack

+++ unit_test/test_files/test_codegen_function.c
@@ -2,7 +2,7 @@ int var = 0;
 
 int main()
 {
-  int a = 0;
+  long a = 0;
   int b = 5;
   a = var;
   var = b + a;
